api/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
werkzeug==3.0.1
APScheduler==3.10.4
python-dotenv==1.0.0
